debugging vertex buffer memory allocation.

+++ source_code/pi_3/mailbox.cpp
@@ -21,12 +21,12 @@ mailbox& mailbox::get_handle()
 
 u32 mailbox::translate_arm_to_vc(void* arm_address)
 {
-    return static_cast<u32>(reinterpret_cast<u64>(arm_address)) | 0xc0000000;
+    return static_cast<u32>(reinterpret_cast<usize>(arm_address)) | 0xc0000000;
 }
 
 void* mailbox::translate_vc_to_arm(u32 vc_address)
 {
-    return reinterpret_cast<void*>(static_cast<u64>(vc_address & ~0xc0000000));
+    return reinterpret_cast<void*>(static_cast<usize>(vc_address & ~0xc0000000));
 }
 
 u32 mailbox::write_read(u32 data, channel a_channel)
